% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{ang_sim_to_deg}
\alias{ang_sim_to_deg}
\title{Convert the angular similarity to degrees.
Function for manual checking and debugging.}
\usage{
ang_sim_to_deg(ang_sim)
}
\arguments{
\item{ang_sim}{Angular similarity.}
}
\value{
scalar. The angle in degrees.
}
\description{
Convert the angular similarity to degrees.
Function for manual checking and debugging.
}
