% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate_biclustering,cadir-method}
\alias{annotate_biclustering,cadir-method}
\title{Annotate the biclustering}
\usage{
\S4method{annotate_biclustering}{cadir}(
  obj,
  universe,
  org,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  ...
)
}
\description{
Annotate the biclustering
}
