% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate_biclusters}
\alias{annotate_biclusters}
\alias{annotate_biclusters,cadir-method}
\title{Perform gene overrepresentation analysis and annotate biclusters.}
\usage{
annotate_biclusters(
  obj,
  universe,
  org,
  caobj = NULL,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  verbose = TRUE,
  method = "goa",
  filter_literature = FALSE,
  cost = c("pval", "logpval", "NES")
)

\S4method{annotate_biclusters}{cadir}(
  obj,
  universe,
  org,
  caobj = NULL,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  verbose = TRUE,
  method = "goa",
  filter_literature = FALSE,
  cost = c("pval", "logpval", "NES")
)
}
\arguments{
\item{obj}{A cadir object.}

\item{universe}{All genes in data set.}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{caobj}{A CA object as obtained from APL.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{alpha}{p-value cutoff for annotating clusters when using cost "pval" or "logpval"}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{verbose}{Toggles verbosity of warnings.}

\item{method}{Either "goa" for gene overrepresentation analysis or "gsea"
for gene set enrichment analysis.}

\item{filter_literature}{Filter out gene sets
named after publications.}

\item{cost}{value that should be
used for constructing the assignment cost matrix.
Can bei either "pval" (adjusted p-value),
"logpval" (adjusted log10(p-value)) or "NES" (normalized enrichment score,
only applicable for GSEA)}
}
\value{
An object of type \code{cadir} with annotated biclusters.
}
\description{
Wrapper function for \code{goa_per_cluster} and \code{annotate_by_goa}.
}
\details{
\code{annotate_biclusters} performs per cluster GOA with a hypergeometric
and annotates the biclustering results from CAdir.
}
