% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate_by_gse}
\alias{annotate_by_gse}
\title{Annotate CAdir results by gene set enrichment}
\usage{
annotate_by_gse(obj, gse_res, alpha = 0.05, cost = c("logpval", "pval", "NES"))
}
\arguments{
\item{obj}{\code{cadir} object with biclustering results. Alternatively could be
a caclust object.}

\item{gse_res}{List of goa results for each bicluster.}

\item{alpha}{p-value cutoff for annotating clusters when using cost "pval" or "logpval"}

\item{cost}{value that should be
used for constructing the assignment cost matrix.
Can bei either "pval" (adjusted p-value),
"logpval" (adjusted log10(p-value)) or "NES" (normalized enrichment score,
only applicable for GSEA)}
}
\value{
Object of the same type as \code{obj}.
}
\description{
\code{annotate_by_gse} takes a biclustering results
and annotates it with the gene overrepresentation analysis results (goa).

Conflicts between clusters that have the
same highest ranking gene set are solved
with the Hungarian/Munkres algorithm.
}
