% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{apl_dir_coords}
\alias{apl_dir_coords}
\title{Calculates APL coordinates for points (columns)
and directions of a cadir object.
point into a precomputed APL space.}
\usage{
apl_dir_coords(cadir, caobj, apl_dir, group)
}
\arguments{
\item{cadir}{A cadir object.}

\item{caobj}{A cacomp object.}

\item{apl_dir}{The direction in the original space for which to
compute the APL coordinates.}

\item{group}{A vector of group indices.
(Usually the cluster belonging to \code{apl_dir})}
}
\value{
A list with the columns (points) and directions projected into the APL space.
}
\description{
Calculates APL coordinates for points (columns)
and directions of a cadir object.
point into a precomputed APL space.
}
