% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{apl_model}
\alias{apl_model}
\title{Create a model to project points into an Association Plot}
\usage{
apl_model(coords, direction, group = NULL)
}
\arguments{
\item{coords}{The CA coordinates to use (should be standard or principal coordinates).}

\item{direction}{Normed direction vector of the APL plot (usually same coordinate type as coords).}

\item{group}{A vector of indices which indicate the points
that belong to the cluster. Only needed here to orient the plot.}
}
\value{
A model that can be used to project new points onto the APL plot.
}
\description{
Create a model to project points into an Association Plot
}
