% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{assign_cells}
\alias{assign_cells}
\title{Assigns cells to the closest direction.}
\usage{
assign_cells(cells, directions)
}
\arguments{
\item{cells}{Row-wise matrix of cell coordinates.}

\item{directions}{Row-wise matrix of directions.}
}
\value{
Vector of same length as nrow(cells) with row-indices of the direction the
cell is assigned to.
}
\description{
Assigns cells to the closest direction.
}
