% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{assign_cts_nes}
\alias{assign_cts_nes}
\title{Assign cell types to clusters using the Hungarian algorithm.}
\usage{
assign_cts_nes(gse_res)
}
\arguments{
\item{gse_res}{List of gene set enrichment results for each bicluster.}
}
\value{
A data frame with the assigned cell types and adjusted p-values.
}
\description{
Uses the hungarian algorithm (assignment problem)
to assign a cell type from the gene set overrepresentation
analysis to one (and only one) cluster.
The cost matrix is based on the Normalized Enrichment score.
}
