% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{assign_genes}
\alias{assign_genes}
\title{Assigns genes to cell clusters based on distance to closest line.}
\usage{
assign_genes(caobj, cadir, qcutoff = NULL, coords = "prin")
}
\arguments{
\item{caobj}{A cacomp object.}

\item{cadir}{A cadir object.}

\item{qcutoff}{The quantile cutoff for gene selection.}

\item{coords}{The coordinates to use for gene selection ("std" or "prin").}
}
\value{
Gene clusters.
}
\description{
\code{assign_genes} removes all genes inside of a sphere which's radius
is defined by the quantile cutoff on the norm of the genes (using
the principal coordinates by default).
The remaining genes are then assigned to the closest line using the
standard coordinates.
}
