% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_sub_graph}
\alias{build_sub_graph}
\title{Build a sub-graph that can be joined to a larger graph.
By itself it is a valid graph too.}
\usage{
build_sub_graph(before, after, before_nm = "start", after_nm = "end")
}
\arguments{
\item{before}{A vector (optimally of factors) with all cluster/node assignments.}

\item{after}{A vector of same length and order as \code{before},
with cluster/node assignments after some change.}

\item{before_nm}{The name of the node before the change.}

\item{after_nm}{The name of the node after the change.}
}
\value{
A data.frame with two columns ("from", "to") representing the edges in the graph.
The data frame can be used to generate a graph.
}
\description{
Build a sub-graph that can be joined to a larger graph.
By itself it is a valid graph too.
}
