% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cadir-class}
\alias{cadir-class}
\title{An S4 class for the CA directional clustering.}
\description{
Class to store biclustering by directions results.
}
\section{Slots}{

\describe{
\item{\code{cell_clusters}}{factors. The assigned cell clusters with cell names in
the names attribute.}

\item{\code{gene_clusters}}{factors. The assigned gene clusters with gene names in
the names attribute.}

\item{\code{directions}}{Matrix of directions by which the data was clustered.}

\item{\code{distances}}{Matrix of distances of points to the respective directions.}

\item{\code{parameters}}{List of used parameters and function name with which results
were generated.}
}}

