% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAdir-package.R
\docType{package}
\name{CAdir-package}
\alias{CAdir}
\alias{CAdir-package}
\title{CAdir: Cluster scRNA-seq data by their directions in CA space}
\description{
CAdir description. WIP.
}
\author{
\strong{Maintainer}: Clemens Kohl \email{kohl@molgen.mpg.de}

Authors:
\itemize{
  \item Martin Vingron \email{vingron@molgen.mpg.de}
}

}
\keyword{internal}
