% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cadir_to_biclust}
\alias{cadir_to_biclust}
\title{Converts a \code{cadir} object to \code{Biclust} object}
\usage{
cadir_to_biclust(cadir)
}
\arguments{
\item{cadir}{A cadir object with cell and gene clusters.}
}
\value{
An object of type "Biclust".
}
\description{
Converts a \code{cadir} object to \code{Biclust} object
}
