% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_cadir}
\alias{check_cadir}
\title{Checks if cadir-class was constructed correctly}
\usage{
check_cadir(object)
}
\arguments{
\item{object}{A \code{cadir} object}
}
\value{
If object is a valid cadir object returns TRUE, otherwise the errors.
}
\description{
Checks if cadir-class was constructed correctly
}
