% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_convergence}
\alias{check_convergence}
\title{Checks whether cluster directions have converged.}
\usage{
check_convergence(now, prev, cutoff = 0.001)
}
\arguments{
\item{now}{Current CAdir object.}

\item{prev}{CAdir object from previous iteration.}

\item{cutoff}{Mean angle between directions from different iterations in
degrees below or equal which convergence is reached.}
}
\value{
TRUE or FALSE, depending if convergence criteria is met.
}
\description{
Checks whether cluster directions have converged.
}
