% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cl2nm}
\alias{cl2nm}
\title{Convert numeric clusters into a cluster name.}
\usage{
cl2nm(i)
}
\arguments{
\item{Cluster}{index.}
}
\value{
Adds prefix "clusters_" to the cluster name.
}
\description{
Convert numeric clusters into a cluster name.
}
