% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{cosine}
\alias{cosine}
\title{Calculate the cosine of the angle between two vectors.}
\usage{
cosine(a, b)
}
\arguments{
\item{a}{Vector or matrix of row-vectors.}

\item{b}{Vector or matrix of row-vectors.}
}
\value{
The cosine of the angle between the vectors.
}
\description{
Calculate the cosine of the angle between two vectors.
}
