% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitmerge.R
\name{decide_split}
\alias{decide_split}
\title{Determine if a cluster should be split.}
\usage{
decide_split(directions, cutoff = deg2rad(30))
}
\arguments{
\item{directions}{A row vector matrix of directions.}

\item{cutoff}{The cutoff angle in degrees.}
}
\value{
A logical indicating if the cluster should be split.
}
\description{
Determine if a cluster should be split.
}
