% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{dist_to_line}
\alias{dist_to_line}
\title{Calculate the distance of points to the directions}
\usage{
dist_to_line(points, lines, pnorm, pos_only = TRUE)
}
\arguments{
\item{points}{Row-wise matrix of points.}

\item{lines}{Row-wise matrix of lines/directions.}

\item{pnorm}{Vector of the norm of the points.}
}
\value{
Matrix of distances of points to the directions.
}
\description{
Calculate the distance of points to the directions
}
