% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_apl.R
\name{.add_cluster_info}
\alias{.add_cluster_info}
\title{Adds information on the clusters to the plotting data frame.}
\usage{
.add_cluster_info(
  plot_df,
  cadir,
  cluster = NULL,
  highlight_cluster = FALSE,
  label_genes = FALSE
)
}
\arguments{
\item{plot_df}{The plotting data frame.}

\item{cadir}{A cadir object for which to compute the APL}

\item{cluster}{The cluster (if any) to highlight cells and genes by.}

\item{highlight_cluster}{If TRUE, highlights the points in belonging to
\code{cluster}, unless colour_by_group is TRUE.}

\item{label_genes}{If TRUE, adds text labels for the
\code{ntop} genes per cluster.}
}
\value{
The enhanced plotting data frame.
}
\description{
Adds information on the clusters to the plotting data frame.
}
