% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_apl.R
\name{.add_scales}
\alias{.add_scales}
\title{Adds size and shape scales to cluster APL ggplot.}
\usage{
.add_scales(ggplt, point_size = 1.5, size_factor = 2)
}
\arguments{
\item{ggplt}{A ggplot2 object. It should be the result of ggplot() with
aes() defined but without any geoms_*.}

\item{point_size}{Size of the points (cells).}

\item{size_factor}{Factor by which the genes are
larger/smaller than \code{point_size}.}
}
\value{
ggplot with with size and shape scale added.
}
\description{
Adds size and shape scales to cluster APL ggplot.
}
