% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_clusters.R
\name{.check_dict}
\alias{.check_dict}
\title{Runs an array of quick tests to check if the @dict slot is coherant.}
\usage{
.check_dict(cadir)
}
\arguments{
\item{cadir}{A CAdir object with a @dict slot.}
}
\value{
TRUE if everything is correct in the dict.
}
\description{
Runs an array of quick tests to check if the @dict slot is coherant.
}
