% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_apl.R
\name{.construct_df}
\alias{.construct_df}
\title{Contructs the basic plotting data frame for cluster_apl}
\usage{
.construct_df(cells = NULL, genes = NULL, model = NULL)
}
\arguments{
\item{cells}{Cell coordinates. If NULL cells will be omitted.}

\item{genes}{Gene coordinates. If NULL genes will be omitted.}

\item{model}{The APL model to project cells and genes.}
}
\value{
Data frame with plotting coordinates and basic information about type and
name.
}
\description{
Contructs the basic plotting data frame for cluster_apl
}
