% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_clusters.R
\name{.correct_dict}
\alias{.correct_dict}
\title{Attempts to correct a malformed dict slot.}
\usage{
.correct_dict(cadir)
}
\arguments{
\item{cadir}{A CAdir object with a @dict slot.}
}
\value{
A CAdir object with corrected dict slot.
}
\description{
Attempts to correct a malformed dict slot.
}
