% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_apl.R
\name{.get_plot_angle}
\alias{.get_plot_angle}
\title{Gets angle between plotting two directions.
Returns NA otherwise.}
\usage{
.get_plot_angle(directions)
}
\arguments{
\item{directions}{Row-wise matrix of directions.}
}
\value{
Angle in degrees or NA if != 2 directions.
}
\description{
Gets angle between plotting two directions.
Returns NA otherwise.
}
