% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_apl.R
\name{.highlight_cluster}
\alias{.highlight_cluster}
\title{Adds specific color schemes to highlight clusters.}
\usage{
.highlight_cluster(ggplt, both = FALSE)
}
\arguments{
\item{ggplt}{A ggplot with geoms etc. already added.}

\item{both}{If TRUE highlights the cluster for both cells and genes,
otherwise cells only.}
}
\value{
ggplot with scale_color added.
}
\description{
Adds specific color schemes to highlight clusters.
}
