% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fc2n}
\alias{fc2n}
\title{Convert a factor to a numeric.}
\usage{
fc2n(f)
}
\arguments{
\item{f}{A vector of factors.}
}
\value{
The factors converted to numbers.
}
\description{
Convert a factor to a numeric.
}
\details{
Assumes that the factors are numbers such as "1".
}
