% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_calc.R
\name{get_ang_sim}
\alias{get_ang_sim}
\title{Calculate angular similarity between two vectors.}
\usage{
get_ang_sim(a, b)
}
\arguments{
\item{a}{Vector or matrix of row-vectors.}

\item{b}{Vector or matrix of row-vectors.}
}
\value{
scalar. The angular similarity.
}
\description{
Calculate angular similarity between two vectors.
}
