% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{get_angle}
\alias{get_angle}
\title{Calculate angle between two vectors.}
\usage{
get_angle(a, b)
}
\arguments{
\item{a}{Vector or matrix of row-vectors.}

\item{b}{Vector or matrix of row-vectors.}
}
\value{
scalar. The angle between the vectors in radian.
}
\description{
Calculate angle between two vectors.
}
