% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{get_apl_mergers}
\alias{get_apl_mergers}
\title{Checks if two clusters should be merged based on their angle in APL space.}
\usage{
get_apl_mergers(
  cadir,
  caobj,
  apl_cutoff_reps = 100,
  method = "random",
  counts = NULL,
  apl_quant = 0.99
)
}
\arguments{
\item{cadir}{A cadir object.}

\item{caobj}{A cacomp object.}

\item{apl_cutoff_reps}{The number of repetitions to use.
Should be between 3-10 for permutation, and >=100 for random.}

\item{method}{Method to use for computing the cutoff.
Either "random" or "permutation".}

\item{counts}{The original count matrix that was used to compute the caobj.}

\item{apl_quant}{The quantile to use for the cutoff.}
}
\value{
A matrix with TRUE/FALSE values indicating if two clusters should be merged.
Importantly the function returns as soon as a candidate is found,
the matrix stays the same size however.
}
\description{
Checks if two clusters should be merged based on their angle in APL space.
}
