% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cluster_idxs}
\alias{get_cluster_idxs}
\title{Get indices of cells belonging to a cluster.}
\usage{
get_cluster_idxs(cadir, cluster)
}
\arguments{
\item{cadir}{A CAdir object.}

\item{cluster}{A cluster name.}
}
\value{
Indices of cells that belong to \code{cluster}.
}
\description{
Get indices of cells belonging to a cluster.
}
