% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_std_num}
\alias{get_std_num}
\title{Extracts the cluster number from standard naming.}
\usage{
get_std_num(nm)
}
\arguments{
\item{nm}{A vector of cluster names starting with "cluster_"}
}
\value{
A vector of cluster numbers.
}
\description{
Extracts the cluster number from standard naming.
}
