% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_plots.R
\name{get_x_y_values}
\alias{get_x_y_values}
\title{Get relative x and y values in relation to the plotting panel from a ggplot.
Solution adapted from Anwer by Allan Cameron at:
https://stackoverflow.com/a/60857307/1376616}
\usage{
get_x_y_values(gg_plot)
}
\arguments{
\item{gg_plot}{A ggplot object.}
}
\value{
dataframe with x and y values relative to the plotting panel.
}
\description{
Get relative x and y values in relation to the plotting panel from a ggplot.
Solution adapted from Anwer by Allan Cameron at:
https://stackoverflow.com/a/60857307/1376616
}
