% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{goa_per_cluster}
\alias{goa_per_cluster}
\title{Perform gene set overrepresentation analysis
for each bicluster and annotate cells based on
the best match.}
\usage{
goa_per_cluster(
  cadir,
  universe,
  org,
  set = "CellMarker",
  min_size = 10,
  max_size = 500,
  filter_literature = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cadir}{A biclustering object of class "cadir".}

\item{universe}{All genes in data set.}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"
and "PanglaoDB"}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{filter_literature}{Filter out gene sets
named after publications.}

\item{verbose}{Toggles verbosity of warnings.}
}
\value{
A list contain the goa results for each cluster.
}
\description{
goa_per_cluster loads the required gene set, formats it and performs
gene overrepresentation analysis for each bicluster in the cadir
object.
}
