% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{gsea_per_cluster}
\alias{gsea_per_cluster}
\title{Perform gene set overrepresentation analysis
for each bicluster and annotate cells based on
the best match.}
\usage{
gsea_per_cluster(
  cadir,
  caobj,
  org,
  set = "CellMarker",
  min_size = 10,
  max_size = 500,
  filter_literature = TRUE
)
}
\arguments{
\item{cadir}{A biclustering object of class "cadir".}

\item{caobj}{A CA object as obtained from APL.}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{filter_literature}{Filter out gene sets
named after publications.}
}
\value{
A list contain the goa results for each cluster.
}
\description{
gsea_per_cluster loads the required gene set, formats it and performs
gene overrepresentation analysis for each bicluster in the cadir
object.
}
