% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.empty}
\alias{is.empty}
\title{Check if a variable is empty (length = 0 but not NULL)}
\usage{
is.empty(x)
}
\arguments{
\item{x}{variable to check}
}
\value{
TRUE if variable is empty, FALSE otherwise.
}
\description{
Check if a variable is empty (length = 0 but not NULL)
}
