% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_std_name}
\alias{is_std_name}
\title{Checks if the cluster names confirm to the standard naming.}
\usage{
is_std_name(nm)
}
\arguments{
\item{nm}{The cluster name(s) to check.}
}
\value{
TRUE/FALSE
}
\description{
Checks if the cluster names confirm to the standard naming.
}
