% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_stored}
\alias{is_stored}
\title{Checks if APL S-alpha cutoff is already calculated.}
\usage{
is_stored(cadir, fun_args)
}
\arguments{
\item{cadir}{Cadir object}

\item{fun_args}{Arguments with which the parent function was called.}
}
\description{
Checks if APL S-alpha cutoff is already calculated.
}
