% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_apl.R
\name{is_xaxis}
\alias{is_xaxis}
\title{Checks if a direction is the x-axis (is the APL direction)}
\usage{
is_xaxis(apl_dir)
}
\arguments{
\item{apl_dir}{A single direction.}
}
\value{
TRUE if apl_dir is equivalent to the x-axis within tolerances,
FALSE otherwise.
}
\description{
Checks if a direction is the x-axis (is the APL direction)
}
