% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{kmeanspp_init}
\alias{kmeanspp_init}
\title{Initialize directions by kmeans++ method}
\usage{
kmeanspp_init(points, k)
}
\arguments{
\item{points}{Row-wise matrix of points to be clustered.}

\item{k}{Number of clusters.}
}
\value{
Vector of indices of the initial directions.
}
\description{
Initialize directions by kmeans++ method
}
