% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_iter}
\alias{log_iter}
\title{Log or append cell clusters to the logs.}
\usage{
log_iter(log, cadir, name)
}
\arguments{
\item{log}{A (empty) list of previous iterations.}

\item{cadir}{A CAdir object with cell clusters.}

\item{name}{Name of the iteration}
}
\value{
List of iterations with cell clusters.
}
\description{
Log or append cell clusters to the logs.
}
