% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panglaodb_mm}
\alias{panglaodb_mm}
\title{PanglaoDB for Mouse}
\format{
An object of class \code{data.frame} with 7054 rows and 14 columns.
}
\usage{
panglaodb_mm
}
\description{
All mouse PanglaoDB marker genes obtained from https://panglaodb.se/markers/PanglaoDB_markers_27_Mar_2020.tsv.gz
Only mouse marker genes.
Oscar Franzén, Li-Ming Gan, Johan L M Björkegren, PanglaoDB: a web server for exploration of mouse and human single-cell RNA sequencing data, Database, Volume 2019, 2019, baz046, https://doi.org/10.1093/database/baz046
}
\keyword{datasets}
