% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{per_cluster_gsea}
\alias{per_cluster_gsea}
\title{Perform gene set overrepresentation analysis
for each bicluster and annotate cells based on
the best match.}
\usage{
per_cluster_gsea(
  cadir,
  caobj,
  org,
  set = "CellMarker",
  min_size = 10,
  max_size = 500,
  filter_literature = TRUE
)
}
\arguments{
\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{filter_literature}{Filter out gene sets
named after publications.}

\item{cabic}{A biclustering object of class "caclust"
as obtained from \code{caclust}.}
}
\value{
A list contain the goa results for each cluster.
}
\description{
per_cluster_goa loads the required gene set, formats it and performs
gene overrepresentation analysis for each bicluster in the cabic
object.
}
