% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{perform_gsea}
\alias{perform_gsea}
\title{Perform GSEA using fgsea.}
\usage{
perform_gsea(
  salpha,
  gene_sets,
  min_size = 15,
  max_size = 500,
  set = "CellMarker"
)
}
\arguments{
\item{salpha}{Named vector containing the Salpha gene scores.}

\item{gene_sets}{Named list of gene sets and their genes.}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}
}
\value{
Data frame of enriched gene sets.
}
\description{
Perform GSEA using fgsea.
}
