% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Summarizes the cell clustering results in a single plot.}
\usage{
plot_clusters(
  cadir,
  caobj,
  point_size = 1,
  size_factor = 1,
  show_genes = FALSE,
  label_genes = FALSE,
  ntop = 5,
  text_size = 16
)
}
\arguments{
\item{cadir}{A cadir object with valid cell clustering results.}

\item{caobj}{A cacomp object.}

\item{point_size}{Size of the points (cells).}

\item{size_factor}{Factor by which the genes are
larger/smaller than \code{point_size}.}

\item{show_genes}{If TRUE, genes are plotted.}

\item{label_genes}{If TRUE, adds text labels for the
\code{ntop} genes per cluster.}

\item{ntop}{Number of genes to label if \code{label_genes = TRUE}.}

\item{text_size}{Size of the text in the plot.}
}
\value{
A plot that summarizes the cell clustering results.
}
\description{
Summarizes the cell clustering results in a single plot.
}
