% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ggsankey}
\alias{plot_ggsankey}
\title{Creates a sankey plot of the clustering results with all splits and merges.}
\usage{
plot_ggsankey(cadir, rm_redund = TRUE)
}
\arguments{
\item{cadir}{A cadir object with valid clustering results.}

\item{rm_redund}{If TRUE, only shows an iteration if something changes.}
}
\value{
A sankey plot of the clustering results.
}
\description{
Creates a sankey plot of the clustering results with all splits and merges.
}
