% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_plots.R
\name{plot_sm_graph}
\alias{plot_sm_graph}
\title{Makes a simple plot of the splits and merges.}
\usage{
plot_sm_graph(cadir, rm_redund = TRUE, size = 3, alpha = 1)
}
\arguments{
\item{cadir}{A cadir object with valid clustering results.}

\item{rm_redund}{If TRUE, only shows an iteration if something changes.}

\item{size}{Size of the nodes.}

\item{alpha}{Between 0 and 1. Sets opacity of the nodes.}
}
\value{
A plot that that shows the
splits and merges as decisions in a graph.
}
\description{
Makes a simple plot of the splits and merges.
}
