% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_calc.R
\name{rad_to_ang_sim}
\alias{rad_to_ang_sim}
\title{Convert the angular similarity to degrees.
Function for manual checking and debugging.}
\usage{
rad_to_ang_sim(rad)
}
\arguments{
\item{rad}{The input radians.}
}
\value{
scalar. radians converted to angular similarity.
}
\description{
Convert the angular similarity to degrees.
Function for manual checking and debugging.
}
