% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rand_idx}
\alias{rand_idx}
\title{return k random row indices from a matrix.}
\usage{
rand_idx(points, k)
}
\arguments{
\item{points}{Row-wise matrix of points.}

\item{k}{Number of indices to return.}
}
\value{
Vector of row-indices.
}
\description{
return k random row indices from a matrix.
}
