% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{random_direction_cutoff}
\alias{random_direction_cutoff}
\title{Random direction association plot coordinates}
\usage{
random_direction_cutoff(
  caobj,
  dims = caobj@dims,
  apl_cutoff_reps = 300,
  axis = "cols"
)
}
\arguments{
\item{caobj}{A cacomp object.}

\item{dims}{Number of dimensions to use for the analysis.}

\item{apl_cutoff_reps}{The number of repetitions to use.
Should be between 3-10 for permutation, and >=100 for random.}

\item{axis}{The axis for which the cutoff should be calculated, either "cols" or "rows".}
}
\value{
Numeric vector of cotangent values of length \code{row_num * apl_cutoff_reps}.
}
\description{
Calculates cotangent values for random directions
}
