% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_clusters_old}
\alias{rename_clusters_old}
\title{Rename clusters and directions so that they match.}
\usage{
rename_clusters_old(cadir)
}
\arguments{
\item{cadir}{A cadir object.}
}
\value{
A cadir object with renamed clusters and directions.
}
\description{
Rename clusters and directions so that they match.
}
\details{
rename_clusters takes a cadir object and renames the clusters and directions
such that they are again coherent (e.g. from 1:5).
}
