% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_monoclusters}
\alias{rm_monoclusters}
\title{Remove clusters only consisting of cells/genes}
\usage{
rm_monoclusters(obj, ...)
}
\arguments{
\item{obj}{biclustering results from CAdir.}

\item{...}{further arguments}
}
\value{
caclust/biclust object with monoclusters removed.
}
\description{
Takes a cadir object and removes all clusters that only consist
of cells or genes.
}
