% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{row_norm}
\alias{row_norm}
\title{Calculate the norm of a row-vector.}
\usage{
row_norm(x)
}
\arguments{
\item{x}{Row-vector or a matrix of row-vectors.}
}
\value{
The norm of the vector.
}
\description{
Calculate the norm of a row-vector.
}
