% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{run_hungarian}
\alias{run_hungarian}
\title{Assign cell types to clusters using the Hungarian algorithm.}
\usage{
run_hungarian(gse_res, cost = "pval")
}
\arguments{
\item{gse_res}{List of gene set enrichment results for each bicluster.}

\item{cost}{Column to use for the cost matrix. Can be
"pval", "logpval" or "NES".}
}
\value{
A data frame with the assigned cell types and adjusted p-values.
}
\description{
Uses the hungarian algorithm (assignment problem)
to assign a cell type from the gene set overrepresentation
analysis to one (and only one) cluster.
The cost matrix is based on the adjusted p-value.
}
