% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_color_mpimg}
\alias{scale_color_mpimg}
\title{ggplot2 scale for the MPI colors and extended swatches.}
\usage{
scale_color_mpimg(name = "mpimg", ...)
}
\arguments{
\item{name}{The name of the scale. Either "mpimg" or "mpi_extend".}

\item{...}{Further arguments to ggplot2::discrete_scale.}
}
\description{
ggplot2 scale for the MPI colors and extended swatches.
}
