% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_fill_gradient_mpimg}
\alias{scale_fill_gradient_mpimg}
\title{Continous ggplot2 scale for the MPI colors and extended swatches.}
\usage{
scale_fill_gradient_mpimg(name = "orange", ...)
}
\arguments{
\item{name}{The name of the scale. Either "mpi" or "mpimg".}

\item{...}{Further arguments to ggplot2::discrete_scale.}
}
\description{
Continous ggplot2 scale for the MPI colors and extended swatches.
}
