% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_dict}
\alias{search_dict}
\title{Searches the dict entry for a given cluster index.}
\usage{
search_dict(dict, query)
}
\arguments{
\item{dict}{The dictionary (cadir@dict).}

\item{query}{A string or index corresponding to the name of the directions.}
}
\value{
The name of the corresponding cluster.
}
\description{
Searches the dict entry for a given cluster index.
}
