% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_cadir}
\alias{show_cadir}
\alias{show,cadir-method}
\title{Print cadir object in console.}
\usage{
show_cadir(object)

\S4method{show}{cadir}(object)
}
\arguments{
\item{object}{A cadir object}
}
\description{
Print cadir object in console.
}
