% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{sign_flip}
\alias{sign_flip}
\title{DETERMINE SIGN FOR SVD SINGULAR VECTORS.
https://www.osti.gov/servlets/purl/920802}
\usage{
sign_flip(points, line)
}
\description{
DETERMINE SIGN FOR SVD SINGULAR VECTORS.
https://www.osti.gov/servlets/purl/920802
}
