% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{slope}
\alias{slope}
\title{Calculate the slope of lines/vectors.
Only works in 2D!}
\usage{
slope(lines, dims = 1:2)
}
\arguments{
\item{lines}{Row-wise matrix of lines/vectors.}

\item{dims}{The 2 dimensions which should be used to calculate the slope.}
}
\value{
Vector of slopes.
}
\description{
Calculate the slope of lines/vectors.
Only works in 2D!
}
