% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitmerge.R
\name{split_clusters}
\alias{split_clusters}
\title{Split clusters based on angle.}
\usage{
split_clusters(cadir, caobj, cutoff = 30, min_cells = 5, make_plots = FALSE)
}
\arguments{
\item{cadir}{A \code{cadir} object.}

\item{caobj}{A \code{cacomp} object.}

\item{cutoff}{Degrees. The cutoff angle to split and merge clusters.
If \code{NULL} the cutoff angle is calculated based on the cutoff angle based
on the quantile defined through \code{apl_quant}.}

\item{min_cells}{Minimum number of cells to form a cluster.}

\item{make_plots}{Logical. If \code{TRUE} plots are generated for each}
}
\value{
A \code{cadir} object with split clusters.
}
\description{
Split clusters based on angle.
}
