% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitmerge.R
\name{sub_cluster}
\alias{sub_cluster}
\title{Split a cluster into 2 sub-clusters.}
\usage{
sub_cluster(cadir, points, idx)
}
\arguments{
\item{cadir}{A \code{cadir} object.}

\item{points}{Matrix of points to be clustered.
Points are expected to be rows.}

\item{idx}{The cluster index to split.}
}
\value{
A \code{cadir} object that contains only the two new clusters.
}
\description{
Split a cluster into 2 sub-clusters.
}
