% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{top_genes}
\alias{top_genes}
\title{Determines the genes above with a S-alpha score above 'cutoff'.
Requires a previous call of \code{rank_genes()}.}
\usage{
top_genes(cadir, cutoff = 0)
}
\arguments{
\item{cadir}{A cadir object.}

\item{cutoff}{Cutoff for the S-alpha score.}
}
\value{
A data frame containing the top genes for each cluster in the cadir object.
}
\description{
Determines the genes above with a S-alpha score above 'cutoff'.
Requires a previous call of \code{rank_genes()}.
}
\details{
Previously ranked genes with an S-alpha score above \code{cutoff} are retained.
}
