% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{total_least_squares}
\alias{total_least_squares}
\title{Perform total least squares regression on points.}
\usage{
total_least_squares(points)
}
\arguments{
\item{points}{Row-wise matrix of points.}
}
\value{
Vector of the regression line (the first PC).
}
\description{
Perform total least squares regression on points.
}
