% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{update_line}
\alias{update_line}
\title{Update directions through Total Least Squares Regression}
\usage{
update_line(points, clusters, lines, k)
}
\arguments{
\item{points}{Row-wise matrix of points.}

\item{clusters}{Vector of cluster assignments for the points.}

\item{lines}{Row-wise matrix of lines.}

\item{k}{Number of clusters.}
}
\value{
Matrix of updated directions.
}
\description{
Update directions through Total Least Squares Regression
}
