% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x2f}
\alias{x2f}
\title{Convert a numeric (or character) to a factor}
\usage{
x2f(n, lvls)
}
\arguments{
\item{n}{Numeric or character vector.}

\item{lvls}{The levels for the new factor vector.}
}
\description{
Convert a numeric (or character) to a factor
}
