% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate}
\alias{annotate}
\title{Perform gene overrepresentation analysis and annotate biclusters.}
\usage{
annotate(
  obj,
  universe,
  org,
  caobj = NULL,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  verbose = TRUE,
  method = "goa",
  filter_literature = TRUE,
  cost = c("pval", "logpval", "NES")
)
}
\arguments{
\item{obj}{Either a \code{caclust} or \code{SingleCellExperiment} object.}

\item{universe}{All genes in data set.}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{verbose}{Toggles verbosity of warnings.}

\item{filter_literature}{Filter out gene sets
named after publications.}

\item{cost}{value that should be
used for constructing the assignment cost matrix.
Can bei either "pval" (adjusted p-value),
"logpval" (adjusted log10(p-value)) or "NES" (normalized enrichment score,
only applicable for GSEA)}

\item{...}{Further arguments.}
}
\value{
An object of type \code{caclust} with annotated biclusters.
}
\description{
Wrapper function for \code{per_cluster_goa} and \code{annotate_by_goa}.
}
\details{
\code{annotate_biclustering} performs per cluster GOA with a hypergeometric
and annotates the biclustering results from CAbiNet.
}
