% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_graph}
\alias{build_graph}
\title{Build igraph from a cadir results.}
\usage{
build_graph(cadir, rm_redund = FALSE)
}
\arguments{
\item{cadir}{A cadir object with valid clustering results.}

\item{rm_redund}{If TRUE, removes all clustering iterations where nothing changed (no splits/merges).}
}
\value{
A directed igraph object with all splits and merges.
}
\description{
Build igraph from a cadir results.
}
