% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cluster_apl}
\alias{cluster_apl}
\title{Plot a cluster with the respective direction/line in an APL.}
\usage{
cluster_apl(
  caobj,
  cadir,
  direction,
  group,
  cluster_id = "NA",
  show_points = TRUE,
  show_lines = TRUE,
  plot_group = FALSE
)
}
\arguments{
\item{caobj}{A cacomp object.}

\item{cadir}{A cadir object for which to compute the APL}

\item{direction}{Normed direction vector of the APL plot.}

\item{group}{A vector of indices which indicate the points
that belong to the cluster. Only needed here to orient the plot.}

\item{cluster_id}{The cluster for which to plot the APL.}
}
\value{
An APL plot (ggplot2 object).
}
\description{
Plot a cluster with the respective direction/line in an APL.
}
