% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cluster_apl}
\alias{cluster_apl}
\title{Plot a cluster with the respective direction/line in an APL.}
\usage{
cluster_apl(
  caobj,
  cadir,
  cluster = NULL,
  direction,
  group,
  show_cells = TRUE,
  show_genes = FALSE,
  show_lines = TRUE,
  highlight_cluster = TRUE,
  colour_by_group = FALSE,
  label_genes = FALSE,
  point_size = 1.5
)
}
\arguments{
\item{caobj}{A cacomp object.}

\item{cadir}{A cadir object for which to compute the APL}

\item{cluster}{The cluster (if any) to highlight cells and genes by.}

\item{direction}{Direction of the APL plot.}

\item{group}{Group to determine the correct direction of the APL plot.
Can be the same as the cluster.}

\item{show_cells}{If TRUE, points (cells) are plotted.}

\item{show_genes}{If TRUE, genes are plotted.}

\item{show_lines}{If TRUE, the directions in cadir are plotted.}

\item{highlight_cluster}{If TRUE, highlights the points in belonging to
\code{cluster}, unless colour_by_group is TRUE.}

\item{colour_by_group}{If TRUE highlights cells that belong to \code{group}.}

\item{point_size}{Size of the points (cells).}
}
\value{
An APL plot (ggplot2 object).
}
\description{
Plot a cluster with the respective direction/line in an APL.
}
