% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cluster_apl}
\alias{cluster_apl}
\title{Plot a cluster with the respective direction/line in an APL.}
\usage{
cluster_apl(
  caobj,
  cadir,
  cluster,
  cluster_id = "NA",
  show_cells = TRUE,
  show_genes = FALSE,
  show_lines = TRUE,
  plot_group = FALSE,
  point_size = 1.5
)
}
\arguments{
\item{caobj}{A cacomp object.}

\item{cadir}{A cadir object for which to compute the APL}

\item{cluster_id}{The name of the cluster for the plot title.}

\item{show_lines}{If TRUE, the directions in cadir are plotted.}

\item{plot_group}{If TRUE, highlights the points in \code{group}.}

\item{point_size}{Size of the points (cells).}

\item{show_points}{If TRUE, points (cells) are plotted.}
}
\value{
An APL plot (ggplot2 object).
}
\description{
Plot a cluster with the respective direction/line in an APL.
}
