% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirclust.R
\name{dirclust}
\alias{dirclust}
\title{Basic clustering by directions in CA space}
\usage{
dirclust(
  points,
  k,
  epochs = 10,
  init = "kmeanspp",
  lines = NULL,
  log = FALSE,
  cadir = NULL
)
}
\arguments{
\item{points}{Matrix of points to be clustered.
Points are expected to be rows.}

\item{k}{Number of clusters.}

\item{epochs}{Number of iterations.}

\item{init}{Initialization method for the lines. Options are 'rand' and
'kmeanspp'.}

\item{lines}{Optional. Row-wise matrix of lines.
If not NULL, then the provided lines are used for initialization.}

\item{log}{Logical. If TRUE, then the function returns a list with
the computed distances and directions at each step.}

\item{cadir}{If a cadir object is provided, uses this instead of \code{lines} to
initialize the clustering.}
}
\value{
An element of class cadir.
}
\description{
Basic clustering by directions in CA space
}
