% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitmerge.R
\name{dirclust_splitmerge}
\alias{dirclust_splitmerge}
\title{Perform Clustering by CA directions with splitting and merging.}
\usage{
dirclust_splitmerge(
  caobj,
  k,
  cutoff = 40,
  min_cells = 5,
  epochs = 15,
  reps = 5,
  make_plots = FALSE
)
}
\arguments{
\item{caobj}{A \code{caclust} object.}

\item{k}{Number of clusters.}

\item{cutoff}{Degrees. The cutoff angle to split and merge clusters.}

\item{min_cells}{Minimum number of cells to form a cluster.}

\item{epochs}{Number of iterations.}

\item{reps}{Number of repetitions to perform the splitting and merging.}

\item{make_plots}{Logical. If \code{TRUE} plots are generated for each
split and merge}
}
\value{
A \code{cadir} object with cell clusters.
}
\description{
Perform Clustering by CA directions with splitting and merging.
}
