% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{permutation_cutoff}
\alias{permutation_cutoff}
\title{Calculates permuted association plot coordinates}
\usage{
permutation_cutoff(
  caobj,
  counts,
  apl_cols,
  group = caobj@group,
  dims = caobj@dims,
  reps = 10,
  python = TRUE
)
}
\arguments{
\item{caobj}{A cacomp object.}

\item{counts}{The original count matrix that was used to compute the caobj.}

\item{apl_cols}{A matrix of association plot coordinates.}

\item{group}{A vector of group indices.}

\item{reps}{The number of repetitions to use.
Should be between 3-10 for permutation, and >=100 for random.}

\item{python}{If TRUE, use python for CA SVD.}
}
\value{
List with permuted apl coordinates ("apl_perm") and, a list of saved ca
components ("saved_ca") that allow for quick recomputation of the CA results.
For random_direction_cutoff this saved_ca is empty.
}
\description{
Calculates matrix of apl coordinates when permuting the original data.
}
