% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Summarizes the cell clustering results in a single plot.}
\usage{
plot_clusters(cadir, caobj, point_size = 1)
}
\arguments{
\item{cadir}{A cadir object with valid cell clustering results.}

\item{caobj}{A cacomp object.}

\item{point_size}{Size of the points (cells).}
}
\value{
A plot that summarizes the cell clustering results.
}
\description{
Summarizes the cell clustering results in a single plot.
}
