% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apl.R
\name{random_direction_cutoff}
\alias{random_direction_cutoff}
\title{Random direction association plot coordinates}
\usage{
random_direction_cutoff(caobj, dims = caobj@dims, apl_cutoff_reps = 300)
}
\arguments{
\item{caobj}{A cacomp object.}

\item{dims}{Number of dimensions to use for the analysis.}

\item{apl_cutoff_reps}{The number of repetitions to use.
Should be between 3-10 for permutation, and >=100 for random.}
}
\value{
List with permuted apl coordinates ("apl_perm") and, a list of saved ca
components ("saved_ca") that allow for quick recomputation of the CA results.
For random_direction_cutoff this saved_ca is empty.
}
\description{
Calculates matrix of apl coordinates for random directions
}
