% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{rank_genes}
\alias{rank_genes}
\title{Rank genes based on the S-alpha score.}
\usage{
rank_genes(cadir, caobj)
}
\arguments{
\item{cadir}{A cadir object.}

\item{caobj}{A cacomp object.}
}
\value{
A modified cadir object with the gene ranking
in the 'gene_ranks' slot.
}
\description{
Rank genes based on the S-alpha score.
}
