% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sm_plot}
\alias{sm_plot}
\title{Plots the graph of the clustering splits and merges
and overlays APL plots over the nodes.}
\usage{
sm_plot(cadir, caobj, rm_redund = TRUE)
}
\arguments{
\item{cadir}{A cadir object with valid clustering results.}

\item{caobj}{A cacomp object.}

\item{rm_redund}{If TRUE, only shows an iteration if something changes.}
}
\value{
A ggplot object showing the split-merge graph and APL plots for each cluster.
}
\description{
Plots the graph of the clustering splits and merges
and overlays APL plots over the nodes.
}
