% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitmerge.R
\name{split_clusters}
\alias{split_clusters}
\title{Split clusters based on angle.}
\usage{
split_clusters(
  cadir,
  caobj,
  cutoff = 30,
  method = "random",
  min_cells = 5,
  make_plots = FALSE,
  counts = NULL,
  apl_cutoff_reps = 100,
  apl_quant = 0.99
)
}
\arguments{
\item{cadir}{A \code{cadir} object.}

\item{caobj}{A \code{caclust} object.}

\item{cutoff}{Degrees. The cutoff angle to split and merge clusters.
If \code{NULL} the cutoff angle is calculated based on the cutoff angle based
on the quantile defined through \code{apl_quant}.}

\item{method}{Method to use for computing the cutoff.
Either "random" or "permutation".}

\item{min_cells}{Minimum number of cells to form a cluster.}

\item{make_plots}{Logical. If \code{TRUE} plots are generated for each}

\item{counts}{The original count matrix that was used to compute the caobj.}

\item{apl_cutoff_reps}{The number of repetitions to use.
Should be between 3-10 for permutation, and >=100 for random.}

\item{apl_quant}{The quantile to use for the APL cutoff. Only used
if cutoff is \code{NULL}.
split and merge}
}
\value{
A \code{cadir} object with split clusters.
}
\description{
Split clusters based on angle.
}
